# openapi.yaml
openapi: 3.1.0
info:
  title: FON Banking REST API
  version: 1.0.0
  summary: REST API demonstracione aplikacije za mobilno bankarstvo.
  description: |
    FON Banking API podržava aktivaciju mobilnog uređaja, PIN autentifikaciju,
    pristup računima i karticama, prenose novca, istoriju transakcija, kursnu
    listu i posredovanje prema NBS IPS QR servisu.

    API je obrazovni prototip. Transferi se izvršavaju samo između računa koji
    postoje u lokalnoj bazi i ne predstavljaju stvarno bankarsko poravnanje.

    Sve poslovne rute koriste prefiks `/api/v1` i razmenjuju JSON. Zaštićene
    rute zahtevaju Sanctum token u zaglavlju `Authorization: Bearer <token>` i
    korisnika čiji je status `active`.
  contact:
    name: FON Banking
    url: https://github.com/Nenad005/fon-banking-backend
  license:
    name: Obrazovna upotreba
    url: https://github.com/Nenad005/fon-banking-backend
servers:
  - url: https://fon-banking.duckdns.org
    description: Javno demonstraciono okruženje
  - url: http://{host}:{port}
    description: Lokalni ili LAN razvojni server
    variables:
      host:
        default: localhost
        description: "`localhost` za isti računar ili LAN IP za fizički telefon."
      port:
        default: "8000"
        description: Port Laravel ili Docker servera.
...
paths:
...
  /api/v1/auth/confirm-pin:
    post:
      tags: [Autentifikacija]
      operationId: confirmPin
      summary: Dodatna potvrda PIN-a
      description: |
        Proverava PIN prijavljenog korisnika pre nego što frontend nastavi
        osetljivu operaciju.

        Ruta je ograničena na pet zahteva u minuti po korisniku i IP adresi.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AcceptJson"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmPinRequest"
            example:
              pin: "2410"
      responses:
        "200":
          description: Identitet je potvrđen.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PinConfirmationResponse"
              example:
                status: success
                message: Identitet je potvrđen.
        "401":
          description: Bearer token nedostaje/nije ispravan ili je PIN pogrešan.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: "#/components/schemas/MessageResponse"
                  - $ref: "#/components/schemas/StatusMessage"
              examples:
                unauthenticated:
                  value:
                    message: Unauthenticated.
                wrongPin:
                  value:
                    status: error
                    message: Pogrešan PIN kod.
        "403":
          $ref: "#/components/responses/InactiveAccount"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Laravel Sanctum token
      description: Token dobijen postavljanjem PIN-a ili prijavom.
  parameters:
    AcceptJson:
      name: Accept
      in: header
      required: true
      description: Zahteva da server vrati odgovor u JSON formatu.
      schema:
        type: string
        const: application/json
      example: application/json
  responses:
...
    InactiveAccount:
      description: Token pripada korisniku čiji nalog nema status `active`.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MessageResponse"
          example:
            message: Nalog nije aktivan.
    ValidationError:
      description: Zahtev nije prošao validaciju ili poslovno pravilo.
      content:
        application/json:
          schema:
            anyOf:
              - $ref: "#/components/schemas/ValidationError"
              - $ref: "#/components/schemas/MessageResponse"
          examples:
            validation:
              value:
                message: The given data was invalid.
                errors:
                  field:
                    - The field is required.
            businessRule:
              value:
                message: Zahtev nije dozvoljen.
    TooManyRequests:
      description: Prekoračen je limit broja zahteva.
      headers:
        Retry-After:
          description: Broj sekundi do sledećeg dozvoljenog zahteva.
          required: false
          schema:
            type: integer
            minimum: 0
        X-RateLimit-Limit:
          description: Maksimalan broj zahteva u vremenskom prozoru.
          required: false
          schema:
            type: integer
        X-RateLimit-Remaining:
          description: Preostali broj zahteva u trenutnom prozoru.
          required: false
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MessageResponse"
          example:
            message: Too Many Attempts.
...
  schemas:
...
    ConfirmPinRequest:
      type: object
      additionalProperties: false
      required: [pin]
      properties:
        pin:
          $ref: "#/components/schemas/Pin"
    Pin:
      type: string
      pattern: "^[0-9]{4}$"
      minLength: 4
      maxLength: 4
      writeOnly: true
      description: Četvorocifreni PIN. Šalje se kao string da bi vodeće nule bile sačuvane.
      example: "2410"
...
    PinConfirmationResponse:
      type: object
      additionalProperties: false
      required: [status, message]
      properties:
        status:
          type: string
          const: success
        message:
          type: string
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
      additionalProperties: true
    StatusMessage:
      type: object
      additionalProperties: false
      required: [status, message]
      properties:
        status:
          type: string
          enum: [success, error]
        message:
          type: string
    ValidationError:
      type: object
      additionalProperties: false
      required: [message, errors]
      properties:
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string